Split utils.
