+++ mart/utils/__init__.py
@@ -1,5 +1,7 @@
 from .adapters import *
 from .export import *
+from .instantiators import *
+from .logging_utils import *
 from .monkey_patch import *
 from .pylogger import *
 from .rich_utils import *
