+++ mart/utils/utils.py
@@ -1,72 +1,24 @@
 import os
-import time
 import warnings
 from glob import glob
 from importlib.util import find_spec
 from pathlib import Path
-from typing import Any, Callable, Dict, List, Tuple
+from typing import Callable, Tuple
 
 import hydra
 from hydra.core.hydra_config import HydraConfig
 from omegaconf import DictConfig, OmegaConf
-from pytorch_lightning import Callback
-from pytorch_lightning.loggers import LightningLoggerBase
-from pytorch_lightning.utilities import rank_zero_only
-from pytorch_lightning.utilities.model_summary import summarize
 
-from mart.utils import pylogger, rich_utils
+from . import pylogger, rich_utils
 
 __all__ = [
-    "close_loggers",
     "extras",
     "get_metric_value",
     "get_resume_checkpoint",
-    "instantiate_callbacks",
-    "instantiate_loggers",
-    "log_hyperparameters",
-    "save_file",
     "task_wrapper",
 ]
 
 log = pylogger.get_pylogger(__name__)
-
-
-def task_wrapper(task_func: Callable) -> Callable:
-    """Optional decorator that wraps the task function in extra utilities.
-
-    Makes multirun more resistant to failure.
-
-    Utilities:
-    - Calling the `utils.extras()` before the task is started
-    - Calling the `utils.close_loggers()` after the task is finished
-    - Logging the exception if occurs
-    - Logging the task total execution time
-    - Logging the output dir
-    """
-
-    def wrap(cfg: DictConfig):
-
-        # apply extra utilities
-        extras(cfg)
-
-        # execute the task
-        try:
-            start_time = time.time()
-            metric_dict, object_dict = task_func(cfg=cfg)
-        except Exception as ex:
-            log.exception("")  # save exception to `.log` file
-            raise ex
-        finally:
-            path = Path(cfg.paths.output_dir, "exec_time.log")
-            content = f"'{cfg.task_name}' execution time: {time.time() - start_time} (s)"
-            save_file(path, content)  # save task execution time (even if exception occurs)
-            close_loggers()  # close loggers (even if exception occurs so multirun won't fail)
-
-        log.info(f"Output dir: {cfg.paths.output_dir}")
-
-        return metric_dict, object_dict
-
-    return wrap
 
 
 def extras(cfg: DictConfig) -> None:
@@ -99,91 +51,57 @@ def extras(cfg: DictConfig) -> None:
         rich_utils.print_config_tree(cfg, resolve=True, save_to_file=True)
 
 
-@rank_zero_only
-def save_file(path, content) -> None:
-    """Save file in rank zero mode (only on one process in multi-GPU setup)."""
-    with open(path, "w+") as file:
-        file.write(content)
+def task_wrapper(task_func: Callable) -> Callable:
+    """Optional decorator that controls the failure behavior when executing the task function.
 
+    This wrapper can be used to:
+    - make sure loggers are closed even if the task function raises an exception (prevents multirun failure)
+    - save the exception to a `.log` file
+    - mark the run as failed with a dedicated file in the `logs/` folder (so we can find and rerun it later)
+    - etc. (adjust depending on your needs)
 
-def instantiate_callbacks(callbacks_cfg: DictConfig) -> List[Callback]:
-    """Instantiates callbacks from config."""
-    callbacks: List[Callback] = []
+    Example:
+    ```
+    @utils.task_wrapper
+    def train(cfg: DictConfig) -> Tuple[dict, dict]:
 
-    if not callbacks_cfg:
-        log.warning("Callbacks config is empty.")
-        return callbacks
+        ...
 
-    if not isinstance(callbacks_cfg, DictConfig):
-        raise TypeError("Callbacks config must be a DictConfig!")
-
-    for _, cb_conf in callbacks_cfg.items():
-        if isinstance(cb_conf, DictConfig) and "_target_" in cb_conf:
-            log.info(f"Instantiating callback <{cb_conf._target_}>")
-            callbacks.append(hydra.utils.instantiate(cb_conf))
-
-    return callbacks
-
-
-def instantiate_loggers(logger_cfg: DictConfig) -> List[LightningLoggerBase]:
-    """Instantiates loggers from config."""
-    logger: List[LightningLoggerBase] = []
-
-    if not logger_cfg:
-        log.warning("Logger config is empty.")
-        return logger
-
-    if not isinstance(logger_cfg, DictConfig):
-        raise TypeError("Logger config must be a DictConfig!")
-
-    for _, lg_conf in logger_cfg.items():
-        if isinstance(lg_conf, DictConfig) and "_target_" in lg_conf:
-            log.info(f"Instantiating logger <{lg_conf._target_}>")
-            logger.append(hydra.utils.instantiate(lg_conf))
-
-    return logger
-
-
-@rank_zero_only
-def log_hyperparameters(object_dict: Dict[str, Any]) -> None:
-    """Controls which config parts are saved by lightning loggers.
-
-    Additionally saves:
-    - Number of model parameters
+        return metric_dict, object_dict
+    ```
     """
 
-    hparams = {}
+    def wrap(cfg: DictConfig):
+        # execute the task
+        try:
+            metric_dict, object_dict = task_func(cfg=cfg)
 
-    cfg = object_dict["cfg"]
-    model = object_dict["model"]
-    trainer = object_dict["trainer"]
+        # things to do if exception occurs
+        except Exception as ex:
+            # save exception to `.log` file
+            log.exception("")
 
-    if not trainer.logger:
-        log.warning("Logger not found! Skipping hyperparameter logging...")
-        return
+            # some hyperparameter combinations might be invalid or cause out-of-memory errors
+            # so when using hparam search plugins like Optuna, you might want to disable
+            # raising the below exception to avoid multirun failure
+            raise ex
 
-    hparams["model"] = cfg["model"]
+        # things to always do after either success or exception
+        finally:
+            # display output dir path in terminal
+            log.info(f"Output dir: {cfg.paths.output_dir}")
 
-    # save number of model parameters
-    summary = summarize(model)
+            # always close wandb run (even if exception occurs so multirun won't fail)
+            if find_spec("wandb"):  # check if wandb is installed
+                import wandb
 
-    hparams["model/params/total"] = summary.total_parameters
-    hparams["model/params/trainable"] = summary.trainable_parameters
-    hparams["model/params/non_trainable"] = summary.total_parameters - summary.trainable_parameters
+                if wandb.run:
+                    log.info("Closing wandb!")
+                    wandb.finish()
 
-    hparams["datamodule"] = cfg["datamodule"]
-    hparams["trainer"] = cfg["trainer"]
+        return metric_dict, object_dict
 
-    hparams["callbacks"] = cfg.get("callbacks")
-    hparams["extras"] = cfg.get("extras")
-
-    hparams["task_name"] = cfg.get("task_name")
-    hparams["tags"] = cfg.get("tags")
-    hparams["ckpt_path"] = cfg.get("ckpt_path")
-    hparams["seed"] = cfg.get("seed")
-
-    # send hparams to all loggers
-    trainer.logger.log_hyperparams(hparams)
+    return wrap
 
 
 def get_metric_value(metric_dict: dict, metric_name: str) -> float:
@@ -204,19 +122,6 @@ def get_metric_value(metric_dict: dict, metric_name: str) -> float:
     log.info(f"Retrieved metric value! <{metric_name}={metric_value}>")
 
     return metric_value
-
-
-def close_loggers() -> None:
-    """Makes sure all loggers closed properly (prevents logging failure during multirun)."""
-
-    log.info("Closing loggers...")
-
-    if find_spec("wandb"):  # if wandb is installed
-        import wandb
-
-        if wandb.run:
-            log.info("Closing wandb!")
-            wandb.finish()
 
 
 def get_resume_checkpoint(config: DictConfig) -> Tuple[DictConfig]:
